+++ Sourcery/main.swift
@@ -9,6 +9,7 @@
 import Foundation
 import Commander
 import PathKit
+import Yams
 
 extension Path: ArgumentConvertible {
     public init(parser: ArgumentParser) throws {
@@ -64,6 +65,37 @@ fileprivate enum Validators {
     }
 }
 
+struct Configuration {
+
+    let sources: [Path]
+    let templates: [Path]
+    let output: Path
+    let args: [String: NSObject]
+
+    init(dict: [String: Any]) {
+        let sources = (dict["sources"] as? [String])?.map({ Path($0) }) ?? []
+        let templates = (dict["templates"] as? [String])?.map({ Path($0) }) ?? []
+        let output = (dict["output"] as? String).map({ Path($0) }) ?? "."
+        let args = dict["args"] as? [String: NSObject] ?? [:]
+
+        self.init(sources: sources, templates: templates, output: output, args: args)
+    }
+
+    init(sources: [Path], templates: [Path], output: Path, args: [String: NSObject]) {
+        self.sources = sources
+        self.templates = templates
+        self.output = output
+        self.args = args
+    }
+
+    func validate() {
+        _ = sources.map(Validators.isFileOrDirectory(path:))
+        _ = templates.map(Validators.isFileOrDirectory(path:))
+        _ = Validators.isFileOrDirectory(path:output)
+    }
+
+}
+
 func runCLI() {
     command(
         Flag("watch",
@@ -81,12 +113,22 @@ func runCLI() {
         Argument<CustomArguments>("args", description: "Custom values to pass to templates.")
     ) { watcherEnabled, disableCache, verboseLogging, sources, templates, output, args in
         do {
-            _ = sources.map(Validators.isFileOrDirectory(path:))
-            _ = templates.map(Validators.isFileOrDirectory(path:))
-            _ = Validators.isFileOrDirectory(path:output)
+            let configuration: Configuration
+
+            let yamlPath: Path = ".sourcery.yml"
+            if let yaml = try? Yams.Node(string: yamlPath.read()).flatten(), let dict = yaml as? [String: Any] {
+                configuration = Configuration(dict: dict)
+            } else {
+                configuration = Configuration(sources: sources,
+                                              templates: templates,
+                                              output: output,
+                                              args: args.arguments)
+            }
+
+            configuration.validate()
 
             let start = CFAbsoluteTimeGetCurrent()
-            if let keepAlive = try Sourcery(verbose: verboseLogging, watcherEnabled: watcherEnabled, cacheDisabled: disableCache, arguments: args.arguments).processFiles(sources, usingTemplates: templates, output: output) {
+            if let keepAlive = try Sourcery(verbose: verboseLogging, watcherEnabled: watcherEnabled, cacheDisabled: disableCache, arguments: configuration.args).processFiles(configuration.sources, usingTemplates: configuration.templates, output: configuration.output) {
                 RunLoop.current.run()
                 _ = keepAlive
             } else {